Updated Jetty, better restart handling, update to cartridges v2

+++ stop.sh
@@ -1,4 +1,14 @@
 #!/bin/bash
-cd ${OPENSHIFT_DATA_DIR}jetty
 
-kill `cat jetty.pid`
+# Ensure the previous startup succeeded and this actually exists.
+if [ -d "${OPENSHIFT_DATA_DIR}jetty" ]; then
+    cd ${OPENSHIFT_DATA_DIR}jetty
+
+    # Ensure the process is actually running.
+    if ps -p `cat jetty.pid` > /dev/null
+    then
+       kill `cat jetty.pid`
+    fi
+fi
+
+
